% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_next.R
\name{last_next}
\alias{last_next}
\title{Impute NA values through Last-Next procedure}
\usage{
last_next(x)
}
\arguments{
\item{x}{a vector, data.frame or array with NA values}
}
\description{
Impute NA values through Last-Next procedure
}

