% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LNarray.R
\name{LNarray}
\alias{LNarray}
\title{Apply Last-Next NA imputation procedure to an array of 
arbitrary dimensions}
\usage{
LNarray(a)
}
\arguments{
\item{x}{an array with NA's}
}
\description{
Apply Last-Next NA imputation procedure to an array of 
arbitrary dimensions
}

