% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LNdataframe.R
\name{LNdataframe}
\alias{LNdataframe}
\title{Apply Last-Next NA imputation procedure to a data.frame}
\usage{
LNdataframe(x)
}
\arguments{
\item{x}{a vector with NA's}
}
\description{
Apply Last-Next NA imputation procedure to a data.frame
}

