% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LNvector.R
\name{LNvector}
\alias{LNvector}
\title{Apply Last-Next NA imputation procedure to a vector}
\usage{
LNvector(x)
}
\arguments{
\item{x}{a vector with NA's}
}
\description{
Apply Last-Next NA imputation procedure to a vector
}

